backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.32
aiosqlite==0.19.0
pyjwt==2.8.0
python-dotenv==1.0.0
asyncpg==0.29.0
alembic==1.12.1
uvicorn==0.23.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0